Implement UGS badges as a commitPublisher extension
Implement RUGS support by depending on commitPublisher and using that
framework. This embeds the JAR from the TeamCity installation because
the API is not otherwise exposed. This was advised by Fedor Rumyantsev
in ticket https://teamcity-support.jetbrains.com/hc/requests/4898304

They have a task to move this into the OpenAPI:
https://youtrack.jetbrains.com/issue/TW-80163/Move-Commit-Status-Publisher-API-into-TeamCity-core